rename: App.js components 폴더 안으로 이동
- refs: #21
